clean-ups and merge
